requirements.txt
chainlit
litellm
python-dotenv
websockets
